DW_Back/requirements.txt
appdirs==1.4.3
astroid==1.5.2
Django==1.11.1
django-cors-headers==2.0.2
django-extensions==1.7.9
django-haystack==2.6.1
django-stdimage==2.4.1
djangorestframework==3.5.3
geopy==1.11.0
graphviz==0.7.1
isort==4.2.5
lazy-object-proxy==1.3.1
mccabe==0.6.1
olefile==0.44
packaging==16.8
Pillow==4.1.1
progressbar2==3.20.0
psycopg2==2.7.1
pyparsing==2.2.0
python-utils==2.1.0
pytz==2017.2
requests==2.14.2
six==1.10.0
Whoosh==2.7.4
wrapt==1.10.10
